requirements.txt
alembic==1.8.1
aniso8601==9.0.1
autoenv==1.0.0
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
passlib==1.7.4
pg==0.1
Pillow==9.2.0
postgres==4.0
protobuf==4.21.7
psycopg2-binary==2.9.4
psycopg2-pool==1.1
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dotenv==0.21.0
python-editor==1.0.4
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.42
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
